TEIIDDES-2458 Fix name in source for Sybase
Attempt to set the correct name in source for tables imported from
Sybase. Sybase does not return a list of schemas, but still requires the
fully qualified name to include the owner.

We workaround this by obtaining the metadata for each imported table and
getting the "schema" from there.

In case of any problem, we fall back on the original behaviour.